WIP Refactor Factory
* This commit mainly refactors the registry for the extractors.
Each concrete extractors registers within dynamic initialization phase
(just before main) in some registry. The registry is an unordered map
containing the name of the extractor and a pointer to a static creation
(factory) function. With this it is possible to get an instance of an
extractor just given there name.