WIP Extractor Registry build
* Refactor the Buildsystem for extractor registry. Unfortunately when
using the registry atm. static linking of the sysmap library is not
possible. This should be changed in future progress, since the ability
of static linking is a design decision. For now, we rely on dynamic
linking and building a shared object. Further this commit add some
libraries for linking when building executables.